% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fasta.R
\name{as.fasta}
\alias{as.fasta}
\title{as.fasta}
\usage{
as.fasta(pr2full, pr2clean)
}
\arguments{
\item{pr2clean}{}
}
\description{
Transform output from CheckPrimer to a list that can be write with write.fasta
}
