% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.D_simple_tag.R
\name{call.D_simple_tag}
\alias{call.D_simple_tag}
\title{call.D_simple_tag}
\usage{
call.D_simple_tag(fastqPath, outputFolder, barcode_path, mismatch = TRUE)
}
\arguments{
\item{fastqPath}{A character string containing the path to the fastq file. The extenstion must be .fastq}

\item{outputFolder}{A character string containing the path to the output folder.}

\item{barcode_path}{A character string containing the path to the barcode file. The extension must be a .txt}

\item{mismatch}{A boolean value, if you allow one mismatch or not.}
}
\value{
Nothing for now (but work on the file)
}
\description{
Function that call the simple tag demultiplexing program of the tool PatPil.
}
\details{
call the tools D_simple_tag of the program PatPil.
For each sequence the program check teh begining of the sequence to see at which file it belongs to.
Only the begining of the sequence is checked and hence no primer is needed.
}
\examples{
\dontrun{
  fastqPath<-".../test/R1.fastq"
  outputFolder <- ".../test_D_simple_tag/R1.fastq"
  barcode_path <- ".../test/barcodesV9test.txt"
  call.D_simple_tag(fastqPath,outputFolder,barcode_path)
}
}
