% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.derep.R
\name{call.derep}
\alias{call.derep}
\title{derep}
\usage{
call.derep(pathFolder, between = 2)
}
\arguments{
\item{pathFolder}{A character string representing the path to the working directory.}

\item{within}{A scalar integer representing the threshold number of occurence of each sequences in a file.}
}
\value{
Nothing, works on the files
}
\description{
Dereplicate each of your fasta files.
}
\details{
Dereplicate each files and the second one works one the dereplicated files.
}
\examples{
\dontrun{
pathFolder <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testderep/derep_ech/"
call.derep(pathFolder,5)
}
}
