% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.derep_ech.R
\name{call.derep_ech}
\alias{call.derep_ech}
\title{derep}
\usage{
call.derep_ech(pathFolder, within = 3)
}
\arguments{
\item{pathFolder}{A character string representing the path to the working directory.}

\item{within}{A scalar integer representing the threshold number of occurence of each sequences in a file.}
}
\value{
Nothing, works on the files
}
\description{
Dereplicate each of your fasta files.
}
\details{
Dereplicate each files and the second one works one the dereplicated files.
}
\examples{
\dontrun{
pathFolder <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testderep/"
call.derep_ech(pathFolder,3)


call.derep_ech(pathFolder = "/home/raphael/Documents/Cadavre/eukaryote/Pilonel_P6_18S/removedNtmp/",
within = 1)


}
}
