% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.qualCheck.R
\name{call.qualCheck}
\alias{call.qualCheck}
\title{call.qualCheck}
\usage{
call.qualCheck(fastqPath, outputFasta, t = 0.01, s = 50, m = 60)
}
\arguments{
\item{fastqPath}{A character string containing the path to the fastq file. The extenstion must be .fastq}

\item{outputFasta}{A character string containing the path to the output folder with the namefile. The extension must be .fasta}

\item{t}{A scalar that represent the percentage of error allowed in the sliding window. Default = 0.01.}

\item{s}{A scalar that represent the size of the sliding window. Default = 50.}

\item{m}{A scalar that represent the minimum size of the sequence to be keep. Default = 60.}
}
\value{
Nothing, but it work on the called file.
}
\description{
Function that call the quality check program of the tool PatPil.
}
\details{
See the \url{http://taraoceans.sb-roscoff.fr/EukDiv/} for more informations on how the quality check is performed.

IMPORTANT : if this function is called in an other manners than by the function \link{preTreatment},
then the function append the files if it is called several times.
Hence you should erase your files if your would like to recall the function.
}
\examples{

\dontrun{
  fastqPath<-".../test/R1.fastq"
  outputFasta <- ".../test/R1.fasta"

  fastqPath <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testqualCheck/unknown.fastq"
  outputFasta <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testqualCheck/unknownClean.fasta"


  call.qualCheck(fastqPath,
       outputFasta,
       t = 0.01,
       s = 50,
       m = 60)
}
}
