% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.RemoveNfasta.R
\name{call.RemoveNfasta}
\alias{call.RemoveNfasta}
\title{call.RemoveNfasta}
\usage{
call.RemoveNfasta(fastaPath, outputFasta)
}
\arguments{
\item{fastaPath}{path to the fasta file}

\item{outputFasta}{path to the output fasta file}
}
\value{
Nothing
}
\description{
call RemoveNfasta function of the PatPil program.
}
\details{
This function applies the RemoveNfasta tools of the PatPil program.
It simply removes the sequences that contains at least one nucleotide 'N' .
}
\examples{
\dontrun{
fastaPath <- ".../fastaTest.fa"
outputFasta <- ".../fastaTest_N.fa"
call.RemoveNfasta(path,outputFolder)
}
}
