% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.RemoveNfastq.R
\name{call.RemoveNfastq}
\alias{call.RemoveNfastq}
\title{call RemoveN}
\usage{
call.RemoveNfastq(fastqPath, outputFastq)
}
\arguments{
\item{fastqPath}{path to the fastq file.}

\item{outputFastq}{path to the output fastq file.}
}
\value{
Nothing, but work on the file.
}
\description{
This function calls the RemoveN program of the tool PatPil.
}
\details{
This function applies the RemoveN tools of the PatPil program.
It simply removes the sequences that contains at least one nucleotide 'N'.
}
\examples{
\dontrun{
fastqPath <- ".../fileTest.fastq"
fastqPath <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testRemoveNfastq/unknown.fastq"
outputFastq <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testRemoveNfastq/unknown_N.fastq"
call.RemoveNfastq(fastqPath,outputFastq)
}
}
