% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.rmSmallSeqfastq.R
\name{call.rmSmallSeqfastq}
\alias{call.rmSmallSeqfastq}
\title{call rmSmallSeqfastq}
\usage{
call.rmSmallSeqfastq(fastqPath, outputFastq, sizemin = 80)
}
\arguments{
\item{fastqPath}{A character string that represent the path to the fastq file.}

\item{outputFastq}{A character string that represent the path to the output fastq file.}

\item{sizemin}{A scalar integer to selected the minimum size.}
}
\value{
Nothing
}
\description{
This function calls the rmSmallSeqfastq program of the tool PatPil.
}
\details{
This function applies the rmSmallSeq tools of the PatPil program.
It simply removes the sequences that are smaller than a fixed integer.
}
\examples{
\dontrun{
fastqPath <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testrmSmallSeqfastq/fastqTest.fastq"
outputFastq <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testrmSmallSeqfastq/fastqTest_rm.fastq"
call.rmSmallSeqfastq(fastqPath,outputFastq,100)
}
}
