% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callFlash.R
\name{callFlash}
\alias{callFlash}
\title{callFlash}
\usage{
callFlash(file_path, fastq1, fastq2, m = 10, M = 100, x = 0.25,
  t = 4)
}
\arguments{
\item{file_path}{A character string containing the path to the fastq file. The extenstion must be .fastq}

\item{fastq1}{A character string containing the path to the output folder with the namefile.
The extenstion must be .fastq and contains R1.}

\item{fastq2}{A character string containing the path to the output folder with the namefile.
The extenstion must be .fastq and contains R2.}

\item{m}{A scalar value that represent the minimum required overlap length between two reads to provide a confident overlap.
Default 10.}

\item{M}{A scalar value that represent the maximum overlap length expected in approximately ninety percent of read pairs.
Default 100.}

\item{x}{A scalar value that represent the maximum allowed ratio between the number of mismatched base pairs and the overlap length. Default 0.25.}

\item{t}{A scalar value that represent the number of worker threads. Default t = 4.}
}
\value{
Nothing.
}
\description{
Function that call the FLASH program. \url{https://ccb.jhu.edu/software/FLASH/}.
}
\examples{

\dontrun{
  file_path<-"/home/raphael/Documents/David_Singer_pipeline/Pipeline2/"
  fastq1 <- "16.02.22_David_Singer_PlateVI_V9_R1.fastq"
  fastq2 <- "16.02.22_David_Singer_PlateVI_V9_R2.fastq"
  callFlash(file_path,fastq1,fastq2)
}
}
