% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CheckPrimer}
\alias{CheckPrimer}
\title{CheckPrimer}
\usage{
CheckPrimer(pr2, primer1, primer2, l_min, l_max)
}
\arguments{
\item{pr2}{pr2 data.frame}

\item{primer1}{vector of all possible primer 1 to consider}

\item{primer2}{vector of all possible primer 2 to consider}

\item{l_min}{bla}

\item{l_max}{bla}
}
\description{
blabla
}
