% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dereplicate.R
\name{Dereplicate}
\alias{Dereplicate}
\title{Dereplicate}
\usage{
Dereplicate(pathFolder, within = 3, between = 2)
}
\arguments{
\item{pathFolder}{A character string representing the path to the folder files.}

\item{within}{A scalar integer representing the threshold number of occurence of each sequences in a file.}

\item{between}{A scalar integer representing the threshold number of occurence of each sequences between files.}
}
\value{
nothing but works on the files
}
\description{
Dereplicate your fasta files.
}
\details{
Works in two step. The first one dereplicate each files and the second one works one the dereplicated files.
}
\examples{
\dontrun{
pathFolder <- "/home/raphael/Documents/
PatPilr_source/testPipeline/testpreTreatment/testDereplicate/"
Dereplicate(pathFolder,10,5)
}
}
