% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.flash.R
\name{install.flash}
\alias{install.flash}
\title{Install the program FLASH}
\usage{
install.flash()
}
\value{
Nothing
}
\description{
call the command "make" to compile the program in Linux environement.
}
\details{
FLASH (Fast Length Adjustment of SHort reads) is a very fast
and accurate software tool to merge paired-end reads
from next-generation sequencing experiments.

source : http://ccb.jhu.edu/software/FLASH/
}
