% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PR2V4.R
\name{PR2V4}
\alias{PR2V4}
\title{Title}
\usage{
PR2V4(pathFile, primerForward = "CYGCGGTAATTCCAGCTC",
  primerReverse = "CRAAGAYGATYAGATACCRT", trim = 0, l_min = 100,
  l_max = 500)
}
\arguments{
\item{l_max}{}
}
\description{
Title
}
\examples{
pathfile <- "/home/raphael/Documents/Rpackage/pipelineR/inst/PR2/PR2.fa"
PR2V4(pathfile)
}
