% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerForDemux.R
\name{primerForDemux}
\alias{primerForDemux}
\title{function used in the preTreatment for double tag}
\usage{
primerForDemux(pathFile)
}
\arguments{
\item{pathFile}{the path where the folder on which there is Primer.txt}
}
\value{
Nothing
}
\description{
function used in the preTreatment for double tag
}
\examples{
\dontrun{
 pathFile <- "/home/raphael/Documents/PatPilr_source/testPipeline/"
}

}
