% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimBase.R
\name{trimBase}
\alias{trimBase}
\title{trim reference database}
\usage{
trimBase(fastaPath, outputFasta, primerForward, primerReverse, trim = 0,
  l_min = 100, l_max = 500, keepPrimer = TRUE)
}
\arguments{
\item{fastaPath}{A character string that represent the path to the input file.}

\item{outputFasta}{A character string that represent the path to the output file.}

\item{primerForward}{A character string that represent the forward primer. Uncertain nucleotide allowed.}

\item{primerReverse}{A character string that represent the reverse primer. Uncertain nucleotide allowed.}

\item{trim}{A value that represent the number of nucleotide that you want trim in the primer.}

\item{l_min}{A value that represent the mininum length of the kept sequences.}

\item{l_max}{A value that represent the maximum length of the kept sequences.}

\item{keepPrimer}{A boolean value, if you want to keep the primer?}
}
\description{
In order to get some 100% of match when you assign your sequences, you have to compare with sequences that have the same "configuration".
Hence, you possibly need to trim your reference base with some primers.
}
\details{
This function will extract from the reference base the portion of interested of the sequences. Schematically we have :
\itemize{
\item SeqPart1 - primerForward - Seq - primerReverse - SeqPart2
}
}
\examples{
\dontrun{
fastaPath <- "/home/raphael/Documents/PatPilr_source/
testPipeline/testpreTreatment/testPR2/utax.fasta"
outputFasta <- "/home/raphael/Documents/PatPilr_source/
testPipeline/testpreTreatment/testPR2/utaxCleanKept.fasta"
outputFastaC <- "/home/raphael/Documents/PatPilr_source/
testPipeline/testpreTreatment/testPR2/utaxClean.fasta"
primerForward <- "GTGAATCATCGAATCTTTGAA"
primerReverse <- "GCATATCAATAAGCGGAGGA"
trim <- 0
l_min = 60
l_max = 600
trimBase(fastaPath,outputFasta,primerForward,primerReverse,trim,l_min,l_max,TRUE)
trimBase(fastaPath,outputFastaC,primerForward,primerReverse,trim,l_min,l_max,FALSE)
}
}
