% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimBasePR2.R
\name{trimBasePR2}
\alias{trimBasePR2}
\title{Trim the PR2 reference database}
\usage{
trimBasePR2(pathFile, primerForward = "CYGCGGTAATTCCAGCTC",
  primerReverse = "CRAAGAYGATYAGATACCRT", trim = 0, l_min = 100,
  l_max = 500, keepPrimer = TRUE)
}
\arguments{
\item{pathFile}{A character string that represent the path to file (absolute path needed)}

\item{primerForward}{A character string that represent the forward primer (uncertain nucleotide allowed)}

\item{primerReverse}{A character string that represent the reverse primer (uncertain nucleotide allowed)}

\item{trim}{A value that represent the number of nucleotide that you want trim in the primer}

\item{l_min}{A value that represent the mininum length of the kept sequences.}

\item{l_max}{A value that represent the maximum length of the kept sequences.}

\item{keepPrimer}{A boolean value, if you want to keep the primer?}
}
\value{
nothing but the file is created in pathFile
}
\description{
This function load, trim and write the PR2 database on a selected directory.
}
\details{
This function will extract from the PR2 the portion of interested of the sequences. Schematically we have :
\itemize{
\item SeqPart1 - primerForward - Seq - primerReverse - SeqPart2
}
}
\examples{
\dontrun{
pathfile <- "/home/raphael/Documents/PatPilr_source/
testPipeline/testpreTreatment/testPR2/PR2CleanKept.fasta"
pathfile <- "/home/raphael/Documents/PatPilr_source/
testPipeline/testpreTreatment/testPR2/PR2Clean.fasta"
trimBasePR2(pathfile,keepPrimer = FALSE)
}
}
