% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackageVersion.R
\name{updatePackageVersion}
\alias{updatePackageVersion}
\title{Update the package version}
\usage{

  updatePackageVersion(packageLocation = "/home/raphael/Documents/PatPilr_source/PatPilr")
}
\arguments{
\item{packageLocation}{location of the main folder of the package}
}
\description{
onyl useful for github update, but if you find another application tell me. :-)
}
\examples{
\dontrun{
 updatePackageVersion()
}

}
