% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllPrimer.R
\name{AllPrimer}
\alias{AllPrimer}
\title{Perform All combination}
\usage{
AllPrimer(primer, direction = "none", trim_nuc = 0)
}
\arguments{
\item{primer}{A character string with wobble nucleotide}

\item{direction}{A character string between "none", "forward" or "reverse"}

\item{trim_nuc}{A scalar value representing the number of nucleotide trimmed (in the direction forward or reverse)}
}
\value{
A data.frame containing all combination depending on the wobble. See Details
}
\description{
This function is used to create the primer forward and reverse for the function \link{preTreatment}
}
\details{
Arguments which should be a character string for example : "CATTGAARARTTGVAT"

The character are then transform following the transformation :

R = "AG",
Y = "CT",
S = "GC",
W = "AT",
K = "GT",
M = "AC",
B = "CGT",
D = "AGT",
H = "ACT",
V = "ACG".

Careful about the number of wobble in the primer, the number of possible combinations can become very large.
}
\examples{
primer <- "CYGCGGTAATTCCAGCTC"
primer <- "CRAAGAYGATYAGATSWMBACCRT"
\dontrun{
AllPrimer(primer,direction = "forward",trim_nuc = 2)
}
}
