% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fasta.R
\name{as.fasta}
\alias{as.fasta}
\title{as.fasta}
\usage{
as.fasta(pr2full, pr2clean)
}
\arguments{
\item{pr2full}{the full database (as.list) pr2}

\item{pr2clean}{the cleaned database (as.list) pr2}
}
\value{
the cleaned database in the right format
}
\description{
Internal function that is used in \code{\link{trimBasePR2}}.
}
\details{
Transform output from \code{\link{CheckPrimer}} to
a list that can be write with \code{\link[seqinr:write.fasta]{write.fasta}}
}
