% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fasta.R
\name{as.fasta}
\alias{as.fasta}
\title{Create the cleaned database}
\usage{
as.fasta(full, clean)
}
\arguments{
\item{full}{A list that represent the full database.}

\item{clean}{A list that represent the cleaned database. Output from \code{\link{CheckPrimer}}.}
}
\value{
the cleaned database in the right format
}
\description{
Internal function that is used in \code{\link{trimBasePR2}} and \code{\link{trimBase}}.
Transform output from \code{\link{CheckPrimer}} to
a list that can be write with \code{\link[seqinr:write.fasta]{write.fasta}}.
}
