% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.D_double_tag.R
\name{call.D_double_tag}
\alias{call.D_double_tag}
\title{call.D_double_tag}
\usage{
call.D_double_tag(fastq_path, outputFolder, bF, bR, pF, pR,
  mismatch = FALSE)
}
\arguments{
\item{fastq_path}{A character string containing the path to the fastq file. The extenstion must be .fastq}

\item{outputFolder}{A character string containing the path to the output folder.}

\item{bF}{A character string containing the path to the barcode forward file. The extension must be a .txt}

\item{bR}{A character string containing the path to the barcode reverse file. The extension must be a .txt}

\item{pF}{A character string containing the path to the primer forward file. The extension must be a .txt}

\item{pR}{A character string containing the path to the primer reverse file. The extension must be a .txt}

\item{mismatch}{A boolean value, if you allow one mismatch or not.}
}
\value{
Nothing for now (but work on the file)
}
\description{
Function that call the double tag demultiplexing program of the tool PatPil.
}
