% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.D_simple_tag.R
\name{call.D_simple_tag}
\alias{call.D_simple_tag}
\title{call D_simple_tag}
\usage{
call.D_simple_tag(fastqPath, outputFolder, barcodePath, mismatch = TRUE)
}
\arguments{
\item{fastqPath}{A character string containing the path to the fastq file. The extenstion must be .fastq}

\item{outputFolder}{A character string containing the path to the output folder.}

\item{barcodePath}{A character string containing the path to the barcode file. The extension must be a .txt}

\item{mismatch}{A boolean value, if you allow one mismatch or not.}
}
\value{
Nothing, but work on the files
}
\description{
This function calls the simple tag demultiplexing program of the tool PatPil.
}
\details{
The demultiplexing simple tag program will check that we are in the configuration :
\itemize{
\item forwardTag - Seq
}

Hence you might have some primer in your sequences but the program will only cares about the tags.
If you would like to trim your primer see ...


IMPORTANT : if this function is called in an other manners than by the function \link{preTreatment},
then the function append the files.
Hence you should erase your files if your would like to recall the function.
}
\examples{
\dontrun{
  fastqPath<-".../test/R1.fastq"
  outputFolder <- ".../test_D_simple_tag/R1.fastq"
  barcodePath <- ".../test/barcodesV9test.txt"

fastqPath <- "/home/raphael/Documents/PatPilr_source/
testPipeline/testpreTreatment/testDemux/D_simple_tag/extendedFrags.fastq"
outputFolder <- "/home/raphael/Documents/PatPilr_source/
testPipeline/testpreTreatment/testDemux/D_simple_tag/demultiplex/"
barcodePath <- "/home/raphael/Documents/PatPilr_source/
testPipeline/testpreTreatment/testDemux/D_simple_tag/barcodes.txt"

  call.D_simple_tag(fastqPath,outputFolder,barcodePath)
}
}
