% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.derep.R
\name{call.derep}
\alias{call.derep}
\title{call derep}
\usage{
call.derep(pathFolder, between = 2)
}
\arguments{
\item{pathFolder}{A character string representing the path to the working directory.}

\item{between}{A scalar integer representing the threshold number of occurence of each sequences between files.}
}
\value{
Nothing, works on the files
}
\description{
This function call the function derep_ech of the tools PatPil.
}
\details{
It takes the dereplicated files of the directory 'pathFolder' and calculate how many times they appears. See Dereplicate for more infos.
}
\examples{
\dontrun{
pathFolder <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testderep/derep_ech/"
call.derep(pathFolder,5)
}
}
