% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.RemoveNfasta.R
\name{call.RemoveNfasta}
\alias{call.RemoveNfasta}
\title{call RemoveNfasta}
\usage{
call.RemoveNfasta(fasta_path, outputFolder)
}
\arguments{
\item{fasta_path}{path to the fasta file}

\item{outputFolder}{path to the output fasta file}
}
\value{
Nothing
}
\description{
call RemoveNfasta function of the PatPil program.
}
\details{
This function applies the RemoveNfasta tools of the PatPil program.
}
\examples{
\dontrun{
path <- "/home/raphael/Documents/PatPilr_source/fastaTest.fa"
outputFolder <- "/home/raphael/Documents/PatPilr_source/fastaTest_N.fa"
call.RemoveNfasta(path,outputFolder)
}
}
