% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.RemoveNfasta.R
\name{call.RemoveNfasta}
\alias{call.RemoveNfasta}
\title{call RemoveNfasta}
\usage{
call.RemoveNfasta(fastaPath, outputFasta)
}
\arguments{
\item{fastaPath}{path to the fasta file.}

\item{outputFasta}{path to the output fasta file.}
}
\value{
Nothing, but work on the file.
}
\description{
This function calls the RemoveNfasta program of the tool PatPil.
}
\details{
This function applies the RemoveNfasta tools of the PatPil program.
It simply removes the sequences that contains at least one nucleotide 'N'.
}
\examples{
\dontrun{
fastaPath <- ".../fastaTest.fa"
fastaPath <- "/home/raphael/Documents/PatPilr_source/
testPipeline/testpreTreatment/testRemoveNfasta/fastaTest.fasta"
outputFasta <- "/home/raphael/Documents/PatPilr_source/
testPipeline/testpreTreatment/testRemoveNfasta/fastaTest_N.fasta"
call.RemoveNfasta(fastaPath,outputFasta)
}
}
