% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.rmSmallSeqfasta.R
\name{call.rmSmallSeqfasta}
\alias{call.rmSmallSeqfasta}
\title{call.rmSmallSeqfasta}
\usage{
call.rmSmallSeqfasta(fastaPath, outputFasta, sizemin = 80)
}
\arguments{
\item{fastaPath}{path to the fasta file}

\item{outputFasta}{path to the output fasta file}
}
\value{
Nothing
}
\description{
call rmSmallSeqfasta function of the PatPil program.
}
\details{
This function applies the rmSmallSeqfasta tools of the PatPil program.
It simply removes the sequences that are smaller than a fixed integer.
}
\examples{
\dontrun{
fastaPath <- ".../fastaTest.fa"
fastaPath <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testrmSmallSeqfasta/fastaTest.fasta"
outputFasta <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testrmSmallSeqfasta/fastaTest_rm.fasta"
call.rmSmallSeqfasta(fastaPath,outputFasta,100)
}
}
