% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callFlash.R
\name{callFlash}
\alias{callFlash}
\title{call the program FLASH}
\usage{
callFlash(outputFolder, fastq1, fastq2, m = 10, M = 100, x = 0.25,
  t = 4)
}
\arguments{
\item{outputFolder}{A character string containing the path to the output folder.}

\item{fastq1}{A character string containing the path to the R1 file.}

\item{fastq2}{A character string containing the path to the R2 file.}

\item{m}{A scalar value that represent the minimum required overlap length between two reads to provide a confident overlap.
Default 10.}

\item{M}{A scalar value that represent the maximum overlap length expected in approximately ninety percent of read pairs.
Default 100.}

\item{x}{A scalar value that represent the maximum allowed ratio between the number of mismatched base pairs and the overlap length. Default 0.25.}

\item{t}{A scalar value that represent the number of worker threads. Default t = 4.}
}
\value{
Nothing.
}
\description{
Function that call the FLASH program. \url{https://ccb.jhu.edu/software/FLASH/}.
}
\examples{
\dontrun{
  outputFolder <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testFlash/"
  fastq1 <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testFlash/PlateVI_R1.fastq"
  fastq2 <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testFlash/PlateVI_R2.fastq"
  callFlash(outputFolder,fastq1,fastq2)
}
}
