% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dereplicate.R
\name{Dereplicate}
\alias{Dereplicate}
\title{Dereplicate}
\usage{
Dereplicate(pathFolder, within = 3, between = 2)
}
\arguments{
\item{pathFolder}{the path to the folder files}

\item{within}{the number of occurence of each sequences in a file.}

\item{between}{the number of occurence of each sequence between files.}
}
\value{
nothing but works on the files
}
\description{
Dereplicate
}
