% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.pr2.R
\name{install.pr2}
\alias{install.pr2}
\title{install pr2}
\usage{
install.pr2()
}
\value{
Nothing
}
\description{
install the PR2 reference database with the R package \url{https://github.com/pr2database/pr2database}.
}
