% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.pr2.R
\name{install.pr2}
\alias{install.pr2}
\title{install pr2}
\usage{
install.pr2(force = FALSE)
}
\arguments{
\item{force}{force argument of install_github}
}
\value{
Nothing
}
\description{
install pr2
}
\examples{
\dontrun{
# install.packages(devtools)
}

}
