% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2Tofasta.R
\name{pr2Tofasta}
\alias{pr2Tofasta}
\title{pr2Tofasta}
\usage{
pr2Tofasta(file)
}
\arguments{
\item{file}{chemin et nom du fichier ex: "/home/test.fasta"}
}
\value{
a file
}
\description{
pr2Tofasta
}
\examples{
\dontrun{
pr2full <-
pr2Tofasta("/home/raphael/Documents/David_Singer_pipeline/Pipeline2/PR2/DB/pr2Rstudio.fa")
}
}
