% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2Tofasta.R
\name{pr2Tofasta}
\alias{pr2Tofasta}
\title{pr2Tofasta}
\usage{
pr2Tofasta(file)
}
\arguments{
\item{file}{optional path to an output file to write the PR2 as a fasta file.
If ommited the database is only loaded in the working environment.}
}
\value{
a \code{\link{list}} that contains the pr2 full
}
\description{
Function that install and load the PR2 database \url{https://github.com/pr2database/pr2database}
}
\details{
Check \code{\link{trimBasePR2}} if you would like trimmed database PR2.
}
\examples{
\dontrun{
pr2full <-
pr2Tofasta("/home/raphael/Documents/David_Singer_pipeline/Pipeline2/PR2/DB/pr2Rstudio.fa")
}
}
