% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preTreatment.R
\name{preTreatment}
\alias{preTreatment}
\title{preTreatment}
\usage{
preTreatment(pathFolder, sep = "_R", m = 10, M = 100, x = 0.25,
  t = 4, mismatch = FALSE, err = 0.01, slide = 50,
  minlength = 60)
}
\arguments{
\item{pathFolder}{A character string containing the path where the program will work.
It should contains the two .fastq file with R1 and R2 inside the name.}

\item{m}{A scalar value that represent the minimum required overlap length between
two reads to provide a confident overlap.  Default 10.}

\item{M}{A scalar value that represent the maximum overlap length expected
in approximately ninety percent of read pairs. Default 100.}

\item{x}{A scalar value that represent the maximum allowed ratio between the number
of mismatched base pairs and the overlap length. Default 0.25.}

\item{t}{A scalar value that represent the number of worker threads. Default t = 4.}

\item{mismatch}{A boolean value if one mismatch is allowed inside the barcode. Default = FALSE}

\item{err}{A scalar that represent the percentage of error allowed in the sliding window. Default = 0.01.}

\item{slide}{A scalar that represent the size of the sliding window. Default = 50.}

\item{minlength}{A scalar that represent the minimum size of the sequence to be keep. Default = 60.}
}
\value{
Noting.
}
\description{
Function that apply a pre-treatment on the fasq file R1 and R2.
}
\details{
The pathFolder argument should contain : namefile_R1.fastq namefile_r2.fastq barcode.txt. The function create the different folder by its own.

1) The function will firstly apply the FLASH to merge the R1 and R2 fastq file. A folder is create named "merged" with the output of the FLASH programs

2) Secondly the function call the simple demultiplexing function of the tool PatPil. The main folder needs to contains a barcode.txt file such that :

barcode TAB < nameoffile.fq >

A folder is create named "demultiplex" which contains the demultiplexed file.

3) Finally, the function call the qualChek function of the tool PatPil. A folder is create named "clean" which contains the fasta files clean by the TARA quality check \url{http://taraoceans.sb-roscoff.fr/EukDiv/}.
}
\examples{
\dontrun{
pathFolder <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testSimple/"
pathFolder <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testDouble/"
pathFolder <- "/home/raphael/Documents/PatPilr_source/testPipeline/testpreTreatment/testAlready/"
preTreatment(pathFolder,M = 250)
}
}
