% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerForDemux.R
\name{primerForDemux}
\alias{primerForDemux}
\title{primer forward and reverse demultiplexing}
\usage{
primerForDemux(pathFile)
}
\arguments{
\item{pathFile}{A character string that represent path where the folder on which there is primer.txt}
}
\value{
Nothing
}
\description{
Function that create forwardtag.txt and reversetag.txt for the double tag demultiplexing
}
\examples{
\dontrun{
 pathFile <- "/home/raphael/Documents/PatPilr_source/testPipeline/"
}

}
