% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimBase.R
\name{trimBase}
\alias{trimBase}
\title{trim the primer of a reference database}
\usage{
trimBase(pathFile, primerForward = "CYGCGGTAATTCCAGCTC",
  primerReverse = "CRAAGAYGATYAGATACCRT", trim = 0, l_min = 100,
  l_max = 500)
}
\arguments{
\item{pathFile}{path to the file}

\item{trim}{number of nucleotide that you want possibly trim to the primer}

\item{l_min}{min length of the keep sequences}

\item{l_max}{max lenght of the keep sequences}

\item{primer1}{primer forward}

\item{primer2}{primer reverse}
}
\value{
nothing but the file is created in pathFile
}
\description{
trim the primer of a reference database
}
\examples{
\dontrun{
pathfile <- "/home/raphael/Documents/PR2.fa"
PR2V4(pathfile)
}
}
