% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimBasePR2.R
\name{trimBasePR2}
\alias{trimBasePR2}
\title{Trim the PR2 reference database}
\usage{
trimBasePR2(pathFile, primerForward = "CYGCGGTAATTCCAGCTC",
  primerReverse = "CRAAGAYGATYAGATACCRT", trim = 0, l_min = 100,
  l_max = 500)
}
\arguments{
\item{pathFile}{A character string that represent the path to file (absolute path needed)}

\item{primerForward}{A character string that represent the forward primer (uncertain nucleotide allowed)}

\item{primerReverse}{A character string that represent the reverse primer (uncertain nucleotide allowed)}

\item{trim}{A value that represent the number of nucleotide that you want trim in the primer}

\item{l_min}{A value that represent the mininum length of the kept sequences.}

\item{l_max}{A value that represent the maximum length of the kept sequences.}
}
\value{
nothing but the file is created in pathFile
}
\description{
This function load, trim and write the PR2 database on a selected directory.
}
\examples{
\dontrun{
pathfile <- "/home/raphael/Documents/PR2.fa"
trimBasePR2(pathfile)
}
}
